move InsightMeta to separate file
